refactor active job test code into helper
