+++ test/active_job_test_case.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+require 'test_helper'
+require 'test_helpers/active_job_test_helpers'
+
+class ActiveJobTestCase < ActiveJob::TestCase
+  include ActiveJobTestHelpers
+end
